endereco-controller-service-dto-entity-repository-OK

+++ src/main/java/br/com/dbc/vemser/financeiro/service/EnderecoService.java
@@ -1,6 +1,7 @@
 package br.com.dbc.vemser.financeiro.service;
 
 
+import br.com.dbc.vemser.financeiro.dto.ClienteDTO;
 import br.com.dbc.vemser.financeiro.dto.EnderecoCreateDTO;
 import br.com.dbc.vemser.financeiro.dto.EnderecoDTO;
 import br.com.dbc.vemser.financeiro.exception.BancoDeDadosException;
@@ -17,33 +18,68 @@ import java.util.stream.Collectors;
 public class EnderecoService extends Servico {
 
     private final EnderecoRepository enderecoRepository;
+    private final ClienteService clienteService;
 
-    public EnderecoService(EnderecoRepository enderecoRepository, ObjectMapper objectMapper) {
+    public EnderecoService(EnderecoRepository enderecoRepository, ClienteService clienteService, ObjectMapper objectMapper) {
         super(objectMapper);
         this.enderecoRepository = enderecoRepository;
+        this.clienteService = clienteService;
+    }
+
+    public List<EnderecoDTO> listarEnderecos() throws BancoDeDadosException, RegraDeNegocioException {
+        return enderecoRepository.listar().stream()
+                .map(endereco -> objectMapper.convertValue(endereco, EnderecoDTO.class))
+                .collect(Collectors.toList());
+    }
+
+    public List<EnderecoDTO> retornarEnderecosDoCliente(Integer idCliente) throws BancoDeDadosException, RegraDeNegocioException {
+        validarPessoaEndereco(idCliente);
+        return this.enderecoRepository.listarEnderecosPorPessoa(idCliente).stream()
+                .map(endereco -> objectMapper.convertValue(endereco, EnderecoDTO.class))
+                .collect(Collectors.toList());
+    }
+
+    public EnderecoDTO retornarEndereco(Integer idEndereco) throws BancoDeDadosException, RegraDeNegocioException {
+        validarEndereco(idEndereco);
+        return objectMapper.convertValue(this.enderecoRepository.retornarEndereco(idEndereco), EnderecoDTO.class);
     }
 
     public EnderecoDTO adicionarEndereco(EnderecoCreateDTO enderecoCreateDTO) throws BancoDeDadosException, RegraDeNegocioException {
+        validarPessoaEndereco(enderecoCreateDTO.getIdCliente());
+        //Validando se o cliente já possui aquele cep registrado no banco de dados.
+        if(listarEnderecos().stream()
+                .filter(enderecoDTO -> enderecoDTO.getIdCliente().equals(enderecoCreateDTO.getIdCliente()))
+                .anyMatch(enderecoDTO -> enderecoDTO.getCep().equals(enderecoCreateDTO.getCep()))){
+            throw new RegraDeNegocioException("Não é possível criar o mesmo CEP no mesmo cliente!");
+        }
         Endereco endereco = objectMapper.convertValue(enderecoCreateDTO, Endereco.class);
         return objectMapper.convertValue(this.enderecoRepository.adicionar(endereco), EnderecoDTO.class);
     }
 
+    public EnderecoDTO atualizarEndereco(Integer idEndereco, EnderecoCreateDTO enderecoCreateDTO) throws BancoDeDadosException, RegraDeNegocioException {
+        validarEndereco(idEndereco);
+        Endereco endereco = objectMapper.convertValue(enderecoCreateDTO, Endereco.class);
+        return objectMapper.convertValue(enderecoRepository.editar(idEndereco, endereco), EnderecoDTO.class);
+    }
+
     public boolean deletarEndereco(Integer idEndereco) throws BancoDeDadosException, RegraDeNegocioException {
+
+        //Validando onde o cliente deve ter ao menos 1 endereço
+        List<EnderecoDTO> enderecoDTOS = retornarEnderecosDoCliente(retornarEndereco(idEndereco).getIdCliente());
+        if(enderecoDTOS.size() == 1){
+            throw new RegraDeNegocioException("É necessário ter ao menos um endereço!");
+        }
         return this.enderecoRepository.remover(idEndereco);
     }
 
-    public EnderecoDTO alterarEndereco(Integer idEndereco, EnderecoCreateDTO enderecoCreateDTO) throws BancoDeDadosException, RegraDeNegocioException {
-        Endereco endereco = objectMapper.convertValue(enderecoCreateDTO, Endereco.class);
-        return objectMapper.convertValue(this.enderecoRepository.editar(idEndereco, endereco), EnderecoDTO.class);
+    private void validarEndereco(Integer idEndereco) throws BancoDeDadosException, RegraDeNegocioException {
+
+        if(listarEnderecos().stream().noneMatch(endereco -> endereco.getIdEndereco().equals(idEndereco))){
+            throw new RegraDeNegocioException("Endereço não encontrado!");
+        }
     }
 
-    public EnderecoDTO retornarEndereco(Integer idEndereco) throws BancoDeDadosException, RegraDeNegocioException {
-        return objectMapper.convertValue(this.enderecoRepository.retornarEndereco(idEndereco), EnderecoDTO.class);
-    }
-
-    public List<EnderecoDTO> retornarEnderecosDoCliente(Integer idCliente) throws BancoDeDadosException, RegraDeNegocioException {
-        return this.enderecoRepository.listarEnderecosPorPessoa(idCliente).stream()
-                .map(endereco -> objectMapper.convertValue(endereco, EnderecoDTO.class))
-                .collect(Collectors.toList());
+    void validarPessoaEndereco(Integer idCliente) throws BancoDeDadosException, RegraDeNegocioException {
+        clienteService.visualizarCliente(idCliente);
     }
 }
